Split the sizing engine: Machine, Size, Size.IR, Eval.Reference
Telomare.Possible (the 1400-line sizing/eval engine) becomes three
stage modules, and Telomare.PossibleData becomes Telomare.Size.IR:

  Telomare.Machine         the composable small-step abstract machine
                           (basicStep/stuckStep/superStep/... algebras,
                           gate machinery, unannotated term builders)
  Telomare.Size            the totality stage: term3ToUnsizedExpr,
                           refinement-derived InputRestrictions,
                           sizeTermM, evalStaticCheck, plus SizingReport
                           and renderSizingCertificate moved down from
                           the driver so Artifact/Certificate no longer
                           depend on it
  Telomare.Size.IR         the abstract-interpretation IRs (UnsizedExpr,
                           InputSizingExpr, StaticCheckExpr, ...)
  Telomare.Eval.Reference  the reference evaluators (AbstractRunTime
                           instances for CompiledExpr and StuckExpr,
                           basicEval, regularEval, evalPartial); absorbs
                           the live remnant of Telomare.RunTime, which
                           is deleted

Also drops forceSizes/zeroToBranch/testB/pathToBranch (no callers).
All five suites pass; sizing regression constants unchanged.
