+++ test/UDTTests.hs
@@ -13,20 +13,20 @@ import Data.List (isInfixOf)
 import Data.Ratio
 import Debug.Trace
 import NatUDTTests (natUDTTests)
-import Telomare.PrettyPrint
 import qualified System.IO.Strict as Strict
 import Telomare.Error
+import Telomare.Eval (SizingOption (..), compile, compileUnitTest,
+                      compileUnitTestNoAbort, runStaticChecks)
 import Telomare.IR.Base
 import Telomare.IR.Builder
 import Telomare.IR.Core
 import Telomare.IR.Loc
 import Telomare.IR.Surface
 import Telomare.IR.Types
-import Telomare.Eval (SizingOption (..), compile, compileUnitTest,
-                      compileUnitTestNoAbort, runStaticChecks)
 import Telomare.Parse (TelomareParser, parseLongExpr, parsePrelude)
-import Telomare.Possible (SizingSettings (SizingSettings))
+import Telomare.PrettyPrint
 import Telomare.Resolve (process, pruneBindings)
+import Telomare.Size (SizingSettings (SizingSettings))
 
 import Test.Tasty
 import Test.Tasty.HUnit
